requirements.txt
fastapi
uvicorn
pandas
numpy
xgboost
streamlit
streamlit-autorefresh
matplotlib
reportlab
sqlalchemy
psycopg2-binary
requests